fix(dashboards): preserve var-* params across widget-editor round-trip
Devin review: navigating into the widget editor and back hard-coded
search: { mode: 'edit' }, wiping the dashboard's var-* variable selections
so pinned values reset to defaults on return.

The editor renders variables at defaults (it doesn't consume var-*), so the
fix round-trips them through the editor's URL: configureWidget carries the
dashboard's var-* into the editor search (new validateSearch retains them),
and navigateBack restores them via functional search alongside mode:'edit'.
Extract the prefix + pick helper + schema fragment into a shared
lib/dashboard-variables/search-params module (deduping the dashboard route's
local copies) so both routes and the actions context share one owner.
